kronitech: PWR provided to Chromebook
We can now providing power to Chromebook, while lit the monitor, on Kronitech docking.
Limitation: when bus-powered with BizLink's emuhost, it can work normally. But it will
reboot at random points when bus-powered with Chromebook.
By this problem, we thus can not freely plug & unplug the DC Jack to trigger the PR_SWAP events.

Changes:
1. Adding debouncing in DC-IN detection, and call pd_pwr_local_change() to adjust
   power control GPIOs, and perform PR_SWAP PD communications
   Also force trigger an HPD event at another deferred interval
2. Some initial changes on DISCHARGE, partially
3. Set flag: PDO_FIXED_EXTERNAL in pd_src_pdo[], since we are wall-powered
4. New functions for controlling power GPIO pins: set_output_voltage(),
   pd_transition_voltage(), pd_pwr_local_change(), check_pr_role()
5. Adjust DP Alt Mode timing for DP Configure.

BUG=none
BRANCH=kronitech
TEST= manual testing under boards of emuhost and Kronitech docking, and plug & unplug
the DC Jack to switch the power role