dfu: add command to ask device enter dfu mode from application
Since we need to shorten the booting time,we skip the dfu-enter detecting in boot-loader
,and on the other hand,we add dedicated command to jump into dfu mode from  application.
The command is sent through usb console,so we need to enable the usb console function.

And for original billboard class support,we still can re-enumerate our device as an
BILLBOARD device as long as the type-c alt mode fail occurred.

Changes:

1.in system.c and system.h ,add bit mask for enter dfu mode and fw signature check.
2.in board.h,add support for usb console.

BUG=none
BRANCH=master
TEST= local testing with STM32F072B EVB over 10 times,no fails.